:truck: api: Move scoring-degradation-service to certification/scoring
