+++ api/src/certification/scoring/domain/services/scoring-degradation-service.js
@@ -1,7 +1,7 @@
-import { AnswerStatus, AssessmentSimulator } from '../../../../lib/domain/models/index.js';
-import { pickAnswerStatusService } from '../../../../lib/domain/services/pick-answer-status-service.js';
-import { AssessmentSimulatorSingleMeasureStrategy } from '../../../certification/flash-certification/domain/models/AssessmentSimulatorSingleMeasureStrategy.js';
-import { pickChallengeService } from '../../../certification/flash-certification/domain/services/pick-challenge-service.js';
+import { AnswerStatus, AssessmentSimulator } from '../../../../../lib/domain/models/index.js';
+import { pickAnswerStatusService } from '../../../../../lib/domain/services/pick-answer-status-service.js';
+import { AssessmentSimulatorSingleMeasureStrategy } from '../../../flash-certification/domain/models/AssessmentSimulatorSingleMeasureStrategy.js';
+import { pickChallengeService } from '../../../flash-certification/domain/services/pick-challenge-service.js';
 
 const downgradeCapacity = ({
   algorithm,
